organization: rename solvers to physics
