+++ src/physics/elasticity_solver.cpp
