+++ src/physics/elasticity_solver.hpp
