+++ src/physics/operators/nonlinear_solid_operators.cpp
