+++ src/physics/operators/nonlinear_solid_operators.hpp
@@ -16,10 +16,10 @@
 #include <memory>
 
 #include "mfem.hpp"
-#include "solvers/physics_utilities/boundary_condition.hpp"
-#include "solvers/physics_utilities/equation_solver.hpp"
-#include "solvers/physics_utilities/finite_element_state.hpp"
-#include "solvers/physics_utilities/solver_config.hpp"
+#include "physics/utilities/boundary_condition.hpp"
+#include "physics/utilities/equation_solver.hpp"
+#include "physics/utilities/finite_element_state.hpp"
+#include "physics/utilities/solver_config.hpp"
 
 namespace serac {
