+++ src/physics/thermal_solver.cpp
