+++ src/physics/utilities/equation_solver.hpp
@@ -18,7 +18,7 @@
 #include <variant>
 
 #include "mfem.hpp"
-#include "solvers/physics_utilities/solver_config.hpp"
+#include "physics/utilities/solver_config.hpp"
 
 namespace serac {
