+++ tests/serac_dynamic_solver.cpp
@@ -11,7 +11,7 @@
 #include "math/mesh_utils.hpp"
 #include "mfem.hpp"
 #include "serac_config.hpp"
-#include "solvers/nonlinear_solid_solver.hpp"
+#include "physics/nonlinear_solid_solver.hpp"
 
 namespace serac {
