+++ tests/serac_quasistatic_solver.cpp
@@ -12,7 +12,7 @@
 #include "math/mesh_utils.hpp"
 #include "mfem.hpp"
 #include "serac_config.hpp"
-#include "solvers/nonlinear_solid_solver.hpp"
+#include "physics/nonlinear_solid_solver.hpp"
 
 namespace serac {
