+++ tests/serac_thermal_solver.cpp
@@ -12,7 +12,7 @@
 #include "math/mesh_utils.hpp"
 #include "mfem.hpp"
 #include "serac_config.hpp"
-#include "solvers/thermal_solver.hpp"
+#include "physics/thermal_solver.hpp"
 
 namespace serac {
