+++ tests/serac_thermal_structural_solver.cpp
@@ -12,7 +12,7 @@
 #include "math/mesh_utils.hpp"
 #include "mfem.hpp"
 #include "serac_config.hpp"
-#include "solvers/thermal_structural_solver.hpp"
+#include "physics/thermal_structural_solver.hpp"
 
 namespace serac {
